Fix code rot in PhabricatorMediaWikiAuthProvider/OAuthAdapter
When testing the oauth extension in preparation for the upcoming
phabricator upgrade, I noticed and a couple of thigns are no longer
working, probably due to upstream changes that we haven't kept up
with.

1. Saving the OAuth endpoint URL wasn't working - it seems
upstream is converting auth provider configuration changes over
to phabricator "application transactions" framework. A simple
call to $config->save() fixes things up at least temporarily.
This might need further attention when we do the next update.

2. We were using "oob" (out of band) callbacks, rather than
passing the callback url to mediawiki via the user's browser.
Phabricator doesn't really like this arrangement because it
doesn't allow phabricator to pass a csrf token as part of the
callback url which is then passed back to phabricator and
compared against the user's session cookie. We hacked around
this in the past but that hack doesn't seem to be working
anymore.

  Instead of finding another hack I just changed the
oauth adapter to pass the callback url to mediawiki in the
url instead of passing "oob." This seems to work fine as
long as the oauth consumer in mediawiki is configured to
allow it. Specifically there is a checkbox in mediawiki
with the following label:

 > Allow consumer to specify a callback in requests and use
 > "callback" URL above as a required prefix.

I believe that this option needs to be enabled for the new
code to work with mediawiki oauth.

Bug: T368

Change-Id: Ia554f1168e17429536f0052620da89523c4cfc32