Add ProjectBurnupGraphProfileMenuItem
Summary:
Based on the preset for Open Tasks (T127903). This should make
it easier for people to add one, since the burnup feature is
quite hard to discover and also helps encouraging consistency
with regards to the icon, label and tooltip.

Most code based on ProjectOpenTasksProfileMenuItem, except for
tooltip, which is from built-in PhabricatorLinkProfileMenuItem.

Bug: T163848

Reviewers: mmodell, O1 Phabricator + Extensions

Reviewed By: mmodell

Subscribers: mmodell, epriestley

Differential Revision: https://phabricator.wikimedia.org/D627