Add ReleaseDetailsCustomField class
This implements the Release Details template for tasks with the
subtype "release". This template field will be used by the
"deployment blockers" tasks initially. The same technique (and code)
can be reused for other task types in the future when we identify a
need for canned markup on various task types. See T114488 for some
background.