Gerrit: Fix up tags url so that it will strip refs/tags
Summary:
This is a preperation change for tags url being supported in gerrit, see https://gerrit-review.googlesource.com/#/c/112238/

I am going to see if i can backport that to 2.14 if not we will get the support in 2.15. Anyways this fixes the tags url which would not have worked anyways due to it not stopping refs/tags/ from the url if it was there.

Reviewers: mmodell, demon

Reviewed By: mmodell

Differential Revision: https://phabricator.wikimedia.org/D702