Make the date field in deadline tasks accessible to herald
This will make it possible to have a herald rule that sets the
task type based on whether there is a due date or not. If the
due date is set the herald rule will change the task type to
deadline.  When a due date is removed, a separate herald rule will
reset the type to 'task' in order to avoid some ui bugs which occur
when deadline tasks are missing their due date.