Update projects names after renames
Summary: `#acl*operations-team` was renamed to `#acl*sre-team` so we should rename it here as well. In addition, #ops-access-reviews is archived and indicates only #operations should be used; as such I'm renaming it here as well, even though I am not sure the process of creating a private subtask for each #sre-access-requests or #ldap-access-requests is still being done.

Reviewers: O1 phabricator.wikimedia.org, mmodell

Reviewed By: O1 phabricator.wikimedia.org, mmodell

Tags: #wikimedia-phabricator-extensions

Differential Revision: https://phabricator.wikimedia.org/D1157